Midterm Example : estimate 'e' with monte carlo
Hi Mr. Canbula, 
It is not the best but I tried to write the a python script that estimating the Eular's number according to the informations that I have learnt from lessons.
You can share it.So my classmates can see if they want to check and study on it.