GIL bypass added
So this is how I managed to bypass GIL. But there is on thing that this works on Linux. If we try to run the code on Windows it throws an exception that says "OSError: [WinError 87]The parameter is incorrect". Yet the issue that I cannot return value from decorator without changing the guess_the_password function still continues.